refact: place ark api and ark types in single file each
