+++ apps/mobile/constants/ark.ts
@@ -35,3 +35,22 @@ export function getArkServer(
 export function getDefaultArkServer(network: Network): ArkServer | undefined {
   return ARK_SERVERS[network][0]
 }
+
+export const ARK_REFRESH_SUBSYSTEM_KEYWORD = 'refresh'
+export const ARK_LIGHTNING_SUBSYSTEM_KINDS = new Set([
+  'invoice',
+  'offer',
+  'lightning_address'
+])
+export const ARK_COUNTERPARTY_TRUNCATE_CHARS = 8
+export const ARK_STALE_EXIT_SUBSYSTEM_NAME = 'bark.exit'
+export const ARK_STALE_EXIT_SUBSYSTEM_KIND = 'start'
+export const ARK_MUTED_STATUSES = new Set(['failed', 'canceled'])
+export const ARK_LNURL_DETAILS_STALE_MS = 60000
+export const ARK_INVALIDATED_QUERY_KINDS = [
+  'wallet',
+  'balance',
+  'address'
+] as const
+export const BARK_ACCESS_TOKEN_PARAM = 'bark_access_token'
+export const BARK_ACCESS_TOKEN_NETWORK: Network = 'bitcoin'
