+++ apps/mobile/hooks/useArkDeleteAccount.ts
@@ -1,11 +1,10 @@
 import { useQueryClient } from '@tanstack/react-query'
 
 import { releaseArkWallet } from '@/api/ark'
+import { ARK_INVALIDATED_QUERY_KINDS } from '@/constants/ark'
 import { deleteArkDatadir } from '@/storage/arkDatadir'
 import { deleteArkMnemonic } from '@/storage/encrypted'
 import { useArkStore } from '@/store/ark'
-
-const INVALIDATED_QUERY_KINDS = ['wallet', 'balance', 'address'] as const
 
 export function useArkDeleteAccount() {
   const queryClient = useQueryClient()
@@ -23,7 +22,7 @@ export function useArkDeleteAccount() {
     await deleteArkDatadir(accountId)
     removeAccount(accountId)
 
-    for (const kind of INVALIDATED_QUERY_KINDS) {
+    for (const kind of ARK_INVALIDATED_QUERY_KINDS) {
       queryClient.removeQueries({ queryKey: ['ark', kind, accountId] })
     }
   }
