+++ apps/mobile/hooks/useArkSendFeeEstimate.ts
@@ -6,18 +6,12 @@ import {
   estimateArkSendOnchainFee
 } from '@/api/ark'
 import { useArkStore } from '@/store/ark'
-import type { ArkFeeEstimate } from '@/types/models/Ark'
+import type {
+  ArkFeeEstimate,
+  UseArkSendFeeEstimateArgs
+} from '@/types/models/Ark'
 
 import { useArkWallet } from './useArkWallet'
-
-export type ArkSendFeeKind = 'arkoor' | 'lightning' | 'onchain'
-
-type UseArkSendFeeEstimateArgs = {
-  accountId: string | null | undefined
-  kind: ArkSendFeeKind | null
-  amountSats: number
-  bitcoinAddress?: string
-}
 
 export function useArkSendFeeEstimate({
   accountId,
